moved a bunch of methods to the respective model classes
